refactor(app): isolate native action routing
